Handle indirect stores with loadaddr as base
Code in checkDefsAndUses analyses indirect stores and array copies
and tracks the value numbers of the base for the indirect store along
with the value numbers of the base's definitions in
_notOptimizableLocalObjectsValueNumbers.

This is not handling the situations in which a loadaddr appears
directly as the base of an indirect store as opposed to being assigned
to a temporary or other symbol that is then used as the base for the
indirect store.

Revised the handling of a use of a loadaddr directly as the base
to match the situation where the value of the loadaddr is accessed
through another symbol.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>