Don't trust frequency if caller is thunk archetype
Bytecodes in a thunk archeytpe don't get executed in interpreter, thus
the iprofiling data and block frequency are incorrect and can't be used
to scale the bytecode size of the callee. This change will leave the
callee's bytecode unscaled. This won't affect callees that are also
thunk archetypes as they are always-worth-inlining targets. The targets
that get affected are the java methods called from leaf MethodHandles.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>