Mark all ConstPatch entries as referenced in CP
During the constantpool creation process, VM will delete constantpool
entires that are not referenced by any bytecode/cp as an optimization.

This will break the patching logic if any of the patched object is not
referenced. So we mark all entries being patched as referenced to keep
them in the RAM class constantpool.

Signed-off-by: Jack Lu <Jack.S.Lu@ibm.com>