Transform invokeBasic, linkToXXX with known object info
- invokeBasic, linkToStatic and linkToSpecial are transformed to direct
call
- linkToVirtual is transformed to indirect call on a resolved virtual
method

Notice that linkToInterface is supposed to be transformed to a indirect
call on an interface method, but currently interface call in OpenJ9 is
represented with unresolved method, which doesn't store j9method and
itable index, and use cpIndex to retrieve the information, which doesn't
work when the interface method is from an object. Thus, linkToInterface
will be addressed in a different PR later.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>