Add utils to refine MethodHandle methods
The following MethodHandle methods can be refined with known object info
 - MethodHandle.invokeBasic
 - MethodHandle.linkToStatic
 - MethodHandle.linkToSpecial
 - MethodHandle.linkToVirtual
 - MethodHandle.linkToInterface

The refined target type is denoted by their name and is determined by
the object info of their arguments.

Notice that MethodHandle.linkToInterface is not transformed in this
commit because in OpenJ9, interface call is represented as an unresolved
method, whose detailed attributes are to be retrieved from constant pool
with a valid cp index. However, for the refined method, the cp index
doesn't resolve to an interface call, thus more changes are needed to
workaround this limitation. The transformation for `linkToInterface`
will be done separately.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>