Maintain stack for getstatic, invokehandle and invokedynamic
Enable iterate with state for methods with invokehandle/invokedynamic
bytecodes such that we can get object info for MH chain and inline the
chain.

For invokehandle, we need to get object info from static final field to
continue inlining.

Notice that more changes are needed to verify the object from static
final field is not changed.